sorting out the configs of the device monitoring service
